rtpoverice:rtp timestamp rollover

+++ link/librtp/MediaStream.h
@@ -56,7 +56,7 @@ typedef struct _MediaStreamTrack
         uint64_t nLastSendPktTimestamp;
         uint64_t nFirstSendPktTimestamp;
         uint32_t nLastRecvPktTimestamp;
-        uint32_t nMostLastRecvTimeAcc;
+        uint64_t nMostLastRecvTimeAcc;
         MediaConfigSet mediaConfig;
         pjmedia_rtp_session  rtpSession;
         pjmedia_rtcp_session rtcpSession;
